chess_bot/pyproject.toml
[project]
name = "chess-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-chess>=1.999",
]
